libusbgx: Fix compilation without libconfig
After refactorization we lost ability to build libusbgx
without libconfig. Let's restore this.

Basic concept is that the whole code related to libconfig usage
should be under #ifdef or compiled only if libconfig support
is enabled.

To prevent future problems of that kind, let's create a header
which is included when compiling without libconfig. Inside it we
like:
	struct s {
	       .import = usbg_get_config_node_int,
	       .export = usbg_set_config_node_int,
	};

are still valid, but all tries to call this functions directly end up
in compilation error. This helps us identify all pieces of code
which should be placed under suitable #ifdef.

Based on initial pull request "Compilation fixes. #4" by:
Bent Bisballe Nyeng (github: aasimon) <deva@aasimon.org>

Reported-by: Bent Bisballe Nyeng (github: aasimon) <deva@aasimon.org>
